fix: can't stop within `api fast` event
AKA "lua loop callback". See `:h api-fast` in nvim.

Solution: replace or remove functions which cannot be called in an
          `api fast` event/context. Such functions include:

          - Vimscript functions (e.g. `vim.fn.*`)
          - Most api functions (e.g. `vim.api.*`)
          - Any function which directly interacts with the editor
            (e.g. executing a vim command or mapping)

There are still a few functions not replaced or removed by this commit
because they do not have a direct replacement and are not likely to
cause issues (because they are called during launch/teardown, not in the
debug hook). For example, `vim.fn.jobstart()` and `vim.fn.jobstop()`.
These can be replaced in the future if necessary, but will require more
careful consideration.

Fixes: #56