feat: implement DAP `OutputEvent` (#58)
* feat: implement DAP `OutputEvent`

Implement DAP `OutputEvent` by using the `:redir` vim cmd to capture
nvim output.

I believe this is the best solution for handling nvim messages/output
(of which lua output constitutes a subset). `:redir` captures all
output/messages (incl. errors, etc.), output is still displayed as usual
within the debuggee, the redirection cannot be observed or detected
(although a single global vim var is required) and there are few
side-effects.

The main caveat is that `:redir` overrides any existing "redirection"
as only 1 redirection is allowed at a time. This "redirection" is
vim-specific and unrelated to fd redirection. I feel that this command
is more used in vim than lua or nvim, and when it is used, it is mostly
used for debugging purposes.

Another caveat is that vim commands, like vimscript functions, cannot be
executed within "lua loop callbacks" or "fast events". So this will not
work within those contexts.

Alternatives considered:

- External/custom nvim ui handler. Would be optimal but requires an rpc
  connection, is more complicated to setup, and probably has too many
  side-effects (messages will likely be hidden in debuggee, headless
  nvim instance will appear to have a UI to user code).

- Collecting output with `:messages` command. Basically requires
  frequently clearing all messages from nvim (e.g. `:messages clear`)
  which user code might detect or rely on. If you don't clear messages,
  it is difficult to reliably determine which messages are actually new.

- Redefining printing functions like `print()`, `io.write`, etc. Would
  work, but probably wouldn't be able to handle error messages,
  nvim_echo calls, or nvim status messages.

- Capturing or redirecting stdout/stderr of debuggee. This would only
  work when nvim is headless, and osv would probably need to be the one
  to launch the debuggee process (as opposed to the user) in order to
  read from, or redirect, the stdio streams.

Related: #42